requirements.txt
tqdm
dbt-core
dbt-snowflake
snowflake-connector-python
apache-airflow-providers-snowflake